:art: [refactor] Enhance metadata generation for quotes and movers, update titles and descriptions
